cuGraph/requirements.txt
# Extra indexes
--extra-index-url https://download.pytorch.org/whl/cu121
--extra-index-url https://pypi.nvidia.com

# Core Python dependencies
numpy==1.24.4
pandas==2.0.3

# PyTorch (CUDA 12.x version)
torch==2.4.0+cu121

# RAPIDS libraries (cuGraph + dependencies)
cudf-cu12
cugraph-cu12
rmm-cu12
